5: Abstrack DatePicker to standalone file, functioning cancel and save cosplay button
