+++ app/src/main/java/com/example/conquest/Application.kt
@@ -7,9 +7,10 @@ import com.example.conquest.data.database.CosplayDatabase
 class ConQuestApplication : Application() {
     val database: CosplayDatabase by lazy {
         Room.databaseBuilder(
-            applicationContext,
-            CosplayDatabase::class.java,
-            "cosplays_database"
-        ).build()
+                applicationContext,
+                CosplayDatabase::class.java,
+                "cosplays_database"
+            ).fallbackToDestructiveMigration(true)
+        .build()
     }
 }
