+++ app/src/main/java/com/example/conquest/data/database/CosplayDatabase.kt
@@ -7,7 +7,7 @@ import com.example.conquest.data.dao.CosplayDao
 import com.example.conquest.data.entity.Cosplay
 import com.example.conquest.data.entity.DateConverter
 
-@Database(entities = [Cosplay::class], version = 1)
+@Database(entities = [Cosplay::class], version = 2)
 @TypeConverters(DateConverter::class)
 abstract class CosplayDatabase: RoomDatabase() {
     abstract fun cosplayDao(): CosplayDao
