+++ app/src/main/java/com/example/conquest/data/entity/Cosplay.kt
@@ -10,12 +10,11 @@ import java.util.Date
 data class Cosplay(
     @PrimaryKey(autoGenerate = true) val uid: Int,
     @ColumnInfo(name = "in_progress") val inProgress: Boolean,
-    @ColumnInfo(name = "planned") val planned: Boolean,
     @ColumnInfo(name = "name") val name: String,
     @ColumnInfo(name = "series") val series: String,
     @ColumnInfo(name = "initial_date") val initialDate: Date,
     @ColumnInfo(name = "due_date") val dueDate: Date?,
-    @ColumnInfo(name = "budget") val budget: Long?
+    @ColumnInfo(name = "budget") val budget: Double?
     )
 
 class DateConverter {
